Change parse_ledger to return iterator not to always spend Vec memory.
