+++ core/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod datamodel;
 pub mod eval;
+pub mod format;
 pub mod load;
 pub mod repl;
